requirements.txt
aiofiles==22.1.0
aioftp==0.21.3
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
asgiref==3.5.2
async-timeout==4.0.2
asyncua==0.9.97
attrs==22.1.0
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==38.0.1
Flask==2.0.3
Flask-AutoIndex==0.6.6
Flask-Markdown==0.3
Flask-Silk==0.2
flask-sock==0.5.2
Flask-WTF==1.0.1
frozenlist==1.3.1
future==0.18.2
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
motor==2.5.1
multidict==6.0.2
psutil==5.9.2
pycparser==2.21
pycryptodomex==3.15.0
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
simple-websocket==0.8.1
six==1.16.0
sortedcontainers==2.4.0
typing-extensions==4.3.0
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
yarl==1.8.1
zipp==3.8.1
